feat(router): add route guard for document availability
A route guard has been added to limit access to documents that are not
available. This is because whenever a user tries to access a file that does not
exist, the backend will create a document with the requested name. This
could potentially lead to a lot of unwanted documents being created.

I am aware that creating new documents is a feature that is going to need to be
implemented in the future, however, since the scope of this project is to
test the capabilities of the collaborative TipTap editor, and the fact that I
am currently hosting this project on a private VM in the cloud, the guard is
implemented to avoid any accidents that could lead to the creation of unwanted
documents.